refactor: extract select field story values to separate file to hide in UI
